updated transform
updated transform position on piepieces to make it child of the player so it follows correctly. made player able to move as wanted. Only one issue left with "Deselect With Fingers" on pressToPlay GameObject, if activated then the players move seperately (which is wanted) but then the other things gets buggy such as the carddeck and pie pieces UI. if it is deselected then everything else works but the players all move at the same time which is not wanted.